Starting with 2.4, Apache needs to read its configuration file even to report its version and build parameters with the "-V" option, probably because the MPM module can be chosen at run time.
However, the system Apache configuration may not be available to the user
building mod_perl, or it may depend on special environment variables or
the like. Make Apache-Test use the test configuration file and server
root instead for the "-V" invocations, if those are available.

Submitted by: Niko Tyni <ntyni debian.org>


git-svn-id: https://svn.apache.org/repos/asf/perl/Apache-Test/trunk@1495994 13f79535-47bb-0310-9956-ffa450edef68